Move tests as integration tests
