+++ rlu/src/simple_stm/mod.rs
@@ -1,3 +1,6 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 pub mod error;
 pub mod stm;
 pub mod transaction;
